requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.26.2
torch==2.1.2
scikit-learn==1.3.2
ta==0.10.2
matplotlib==3.8.2
tqdm==4.66.1
python-dotenv==1.0.0
joblib==1.3.2 